fix(Discussion): replace links context with stable Link component which gets everything as props
Because e.g. discussion was hoisted we often created a new link component on the fly. By moving
everything to props this becomes unnecessary and it's easy to always provide the same link component
which is stable

BREAKING CHANGE: all discussion context now need to provide Link instead of links