chore(search): Use include_type_name parameter, and align Kibana version in Docker
In [ElasticSearch] 6.8, the index creation, index template, and mapping
APIs support a query string parameter (include_type_name) which
indicates whether requests and responses should include a type name. It
defaults to true, and should be set to an explicit value to prepare to
upgrade to 7.0. Not setting include_type_name will result in a
deprecation warning. Indices which don’t have an explicit type will use
the dummy type name _doc.