feat(admin): get rid of custom express server
getting rid of the express server also required refactoring on how the environment-variables are being loaded,
as previously a own solution was added (backe when next.js didn't support env-variables out of the box).

In order for the styleguide to be able to access the 'SG_' prefixed env-variables, the 'SG_' variables are
manually rexported in the next.js' configs 'env' object.