feat(shop): serialize mail settings into metadata
Introduces a mechanism to prevent the sending of emails by checking
the metadata of a subscription for the republik.payments.mail.settings key.

The value key should be a comma-separated list of mail setting options
with a boolean value.

For example, to turn off cancellation confirmations, use:

confirm:cancel=false

If you want to suppress cancellation confirmations and the ending
notice email, use:

confirm:cancel=false,notice:ended=false

Note that the option keys used here might change in the future.