feat: add view for querying published and prepublished documents
Currently, our database structure makes it hard to efficiently
query both published and prepublished documents.

The new `documents` view offers a fast way to retrieve and filter
all published documents.

Commit content is excluded to keep the table smaller and because
filtering by content is not needed. This can be changed later if
requirements evolve.